+++ test/args-test.cc
@@ -7,7 +7,7 @@
 
 #include "fmt/args.h"
 
-#include "gtest.h"
+#include "gtest/gtest.h"
 
 TEST(args_test, basic) {
   auto store = fmt::dynamic_format_arg_store<fmt::format_context>();
