+++ test/format-test.cc
@@ -30,7 +30,7 @@
 
 #undef index
 
-#include "gmock.h"
+#include "gmock/gmock.h"
 #include "gtest-extra.h"
 #include "mock-allocator.h"
 #include "util.h"
